Disable autocomplete which confuses React
Firefox (but not Chrome) remembers the values of the <input> elements
across a page reload.  This would be useful, but React doesn't know
about the values so the computed attribute value is wrong.

I tried using refs to update state to have the correct value the first
time the component is mounted, but this just resulted in infinite loops.
I think that the ultimate solution will be to make this data props and
stash it in localStorage or some such.